restructure controls
